+++ app/database-models/src/main/scala/com/qwery/database/models/UpdateCount.scala
@@ -2,6 +2,6 @@ package com.qwery.database.models
 
 import com.qwery.database.JSONSupport.JSONProductConversion
 
-case class UpdateCount(count: Int, __id: Option[Int] = None) {
+case class UpdateCount(count: Long, __id: Option[Long] = None) {
   override def toString: String = this.toJSON
 }
